feat: add FetchingMailSettingsOptionsMapper
